Move Layout trait to new module
