+++ crates/kas-core/src/core/mod.rs
@@ -6,6 +6,7 @@
 //! Core widget types
 
 mod data;
+mod layout;
 mod node;
 mod scroll_traits;
 mod widget;
@@ -16,6 +17,7 @@ mod widget_id;
 pub mod impls;
 
 pub use data::*;
+pub use layout::*;
 pub use node::Node;
 pub use scroll_traits::*;
 pub use widget::*;
